Compute growth and decline matrices

DESCRIPTION NEEDED

ComputeGrowthAndDeclineMatrices2(
  growthIncrements,
  ages,
  gcids,
  pools,
  rootParameters,
  turnoverParams,
  biomassToCarbonRate,
  swMult = 1,
  hwMult = 1
)

Arguments

growthIncrements

a hash table of growth increments by gcid, by age

ages

the stand age, (the inventory age) stands will be simulated to this age in the final pass

gcids

integer vector of growth curve ids

pools

Matrix (DESCRIPTION NEEDED)

rootParameters

data.frame specifying root parameters

turnoverParams

`data.frame specifying turnover parameters

biomassToCarbonRate

numeric DESCRIPTION NEEDED

swMult

softwood multiplier? (DESCRITION NEEDED)

hwMult

hardwood multiplier? (DESCRITION NEEDED)