Adjust cohort ages

Adjust cohort ages to represent a year before or after a given date. Optionally include disturbance events that reset ages to year 0.

adjustCohortAges(
  cohortAges,
  yearInput,
  yearOutput,
  disturbanceEvents = NULL,
  delay = NULL,
  default = NULL,
  warn = TRUE
)

Arguments

cohortAges

data.table. Table of cohort ages with an ID column and the numeric columns 'age'. Optionally include a 'delay' column with regeneration delays for each cohort. A standard delay can also be provided with the delay argument. Optionally include a 'default' column with the default age for the cohort if it cannot be otherwise calculated. A default age can also be provided with the default argument.

yearInput

integer. The year that the 'ages' column in cohortAges represents.

yearOutput

integer. The year that cohort ages must be adjusted to.

disturbanceEvents

data.table. Optional. Table of disturbance events with the ID column in cohortAges and another column 'year' of when the disturbance occurred. This can also include rows for years that cohorts were first established.

delay

integer. Optional. Regeneration delay after a disturbance event.

default

integer. A default age for cohorts is otherwise unknown. If yearOutput precedes yearInput, the age cannot be calculated for cohorts that have disturbances between yearOutput and yearInput but none before yearOutput. default will be assigned in these cases. NOTE: if a cohort has an age lesser than the difference in these years, it is assumed that a disturbance event must have occurred when age == 0. default will be assigned in these cases.

warn

logical. Warn if ages cannot be calculated or if the provided disturbances do not match the input data.

Value

cohortAges with ages adjusted to yearOutput.