Verify and check transactions

Post-simulations, this function takes the last two years of simulations, and verifies that the C transfers are as expected on n random pixelGroups. It randomly selects three pixelGroups, calculated the expected transactions from time end(sim) and time end(sim) - 1, and compares them to the last year of simulations. Cases of disturbed, non disturbed, sw, hw, young and old stands were tested.

checkTransactions(
  cbmPools,
  pixelKeep,
  opMatrixCBM,
  allProcesses,
  pixelGroupC,
  pooldef,
  timeSim,
  outputsSim,
  n = 3
)

Arguments

cbmPools

from sim$cbmPools

pixelKeep

from sim$pixelKeep

opMatrixCBM

from sim$opMatrixCBM

allProcesses

from sim$allProcesses

pixelGroupC

from sim$pixelGroupC

pooldef

from sim$pooldef

timeSim

from time(sim)

outputsSim

from outputs(sim)

n

number of pixelGroups (i.e., size of subset) to check

Structure

Read-in data

randomly select 3 pixelGroups

read-in carbon pools at timeSim and timeSim - 1 for the three pixelGroup

pull-out matrices applied to these three pixelGroups between t0 and t1

matrices into a data.table

prep pools at t0

merge three pools at t0 with their respective transaction matrices ids

apply all matrices (8.1:8.9)

TODO: I am passing a whole simList here for simplification, but only these objects needed:

$cbmPools;

$pixelKeep;

$opMatrixCBM;

$allProcesses;

$pooldef.

TODO: there are still a bunch of checks that need to be changed to assertions (if statements?)

All arguments come from a simList object of the same name, except timeSim (which comes from time(sim)) and outputsSim (which comes from outputs(sim)).