Species match

Retrieve species metadata by matching species names or other identifiers with columns in sppEquivalencies.

sppMatch(
  species,
  match = c("LandR", "Latin_full", "EN_generic_short", "EN_generic_full"),
  otherNames = NULL,
  return = NULL,
  checkNA = !is.null(return),
  sppEquivalencies = NULL
)

Arguments

species

Species identifiers.

match

character. sppEquivalencies column(s) to match species with.

otherNames

list. A list of alternative species identified to allow in matching. Item names must match species and item contents must be vectors of additional allowable matches.

return

character. sppEquivalencies columns to return. All columns will be returned by default.

checkNA

logical. Check for NA values in the returned columns. Defaults to TRUE if the return argument is used; otherwise FALSE.

sppEquivalencies

data.table. Table of species identifiers and metadata. Defaults to LandR::sppEquivalencies_CA.

Value

data.table. Subset of sppEquivalencies with 1 row per species.