CBM-CFS3 Disturbances Match

Match disturbance names with CBM-CFS3 spatial unit disturbances.

spuDistMatch(
  distTable,
  ask = interactive(),
  nearMatches = TRUE,
  listDist = NULL,
  ...
)

Arguments

distTable

data.table with columns 'spatial_unit_id' and 'name' (or 'distName'). The name column will be matched with disturbance names and descriptions in the CBM-CFS3 database.

ask

logical. If TRUE, prompt the user to choose the correct disturbance matches. If FALSE, the function will look for exact name matches.

nearMatches

logical. Allow for near matches; e.g. "clearcut" can match "clear-cut".

listDist

data.table. Optional. Result of a call to spuDist. A list of possible disturbances in the spatial unit(s) with columns 'spatial_unit_id', 'disturbance_type_id', 'disturbance_matrix_id', 'name', 'description'.

...

arguments to spuDist for listing the possible disturbances in the spatial units.

Value

data.table with columns 'spatial_unit_id' 'disturbance_type_id', 'disturbance_matrix_id', 'name', 'description'